Simplifying code
