+++ compiler/src/test/kotlin/land/sungbin/composeinvestigator/compiler/_compilation/KotlinK2Compiler.kt
@@ -7,15 +7,25 @@
 
 package land.sungbin.composeinvestigator.compiler._compilation
 
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.TextRange
+import com.intellij.openapi.util.text.StringUtilRt
 import com.intellij.openapi.vfs.StandardFileSystems
 import com.intellij.openapi.vfs.VirtualFileManager
+import com.intellij.psi.PsiFileFactory
+import com.intellij.psi.impl.PsiFileFactoryImpl
 import com.intellij.psi.search.ProjectScope
+import com.intellij.testFramework.LightVirtualFile
+import java.nio.charset.StandardCharsets
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.jvm.JvmIrDeserializerImpl
-import org.jetbrains.kotlin.cli.common.fir.FirDiagnosticsCompilerResultsReporter
 import org.jetbrains.kotlin.cli.common.messages.AnalyzerWithCompilerReport
+import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
+import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSourceLocation
+import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.cli.common.messages.MessageRenderer
+import org.jetbrains.kotlin.cli.jvm.compiler.EnvironmentConfigFiles
 import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
 import org.jetbrains.kotlin.cli.jvm.compiler.PsiBasedProjectFileSearchScope
 import org.jetbrains.kotlin.cli.jvm.compiler.TopDownAnalyzerFacadeForJVM
@@ -23,8 +33,11 @@ import org.jetbrains.kotlin.cli.jvm.compiler.VfsBasedProjectEnvironment
 import org.jetbrains.kotlin.cli.jvm.compiler.pipeline.convertToIrAndActualizeForJvm
 import org.jetbrains.kotlin.config.CommonConfigurationKeys
 import org.jetbrains.kotlin.config.CompilerConfiguration
+import org.jetbrains.kotlin.config.IrVerificationMode
 import org.jetbrains.kotlin.config.JVMConfigurationKeys
+import org.jetbrains.kotlin.config.JvmTarget
 import org.jetbrains.kotlin.config.languageVersionSettings
+import org.jetbrains.kotlin.config.messageCollector
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporterFactory
 import org.jetbrains.kotlin.diagnostics.impl.BaseDiagnosticsCollector
 import org.jetbrains.kotlin.diagnostics.rendering.RootDiagnosticRendererFactory
@@ -41,22 +54,63 @@ import org.jetbrains.kotlin.fir.pipeline.FirResult
 import org.jetbrains.kotlin.fir.pipeline.buildResolveAndCheckFirFromKtFiles
 import org.jetbrains.kotlin.fir.session.FirJvmSessionFactory
 import org.jetbrains.kotlin.fir.session.environment.AbstractProjectEnvironment
-import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
+import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.kotlin.name.Name
 import org.jetbrains.kotlin.platform.CommonPlatforms
 import org.jetbrains.kotlin.platform.jvm.JvmPlatforms
+import org.jetbrains.kotlin.psi.KtFile
+import org.jetbrains.kotlin.resolve.AnalyzingUtils
+
+class SourceFile(
+  val name: String,
+  val source: String,
+  private val ignoreParseErrors: Boolean = false,
+  val path: String = "",
+) {
+  fun toKtFile(project: Project): KtFile {
+    val shortName = name.substring(name.lastIndexOf('/') + 1).let { name ->
+      name.substring(name.lastIndexOf('\\') + 1)
+    }
+
+    val virtualFile = object : LightVirtualFile(
+      /* name = */ shortName,
+      /* language = */ KotlinLanguage.INSTANCE,
+      /* text = */ StringUtilRt.convertLineSeparators(source),
+    ) {
+      override fun getPath(): String = "${this@SourceFile.path}/$name"
+    }
+    virtualFile.charset = StandardCharsets.UTF_8
+
+    val factory = PsiFileFactory.getInstance(project) as PsiFileFactoryImpl
+    val ktFile = factory.trySetupPsiForFile(
+      /* lightVirtualFile = */ virtualFile,
+      /* language = */ KotlinLanguage.INSTANCE,
+      /* physical = */ true,
+      /* markAsCopy = */ false,
+    ) as KtFile
+
+    if (!ignoreParseErrors) AnalyzingUtils.checkForSyntacticErrors(ktFile)
+    return ktFile
+  }
+}
+
+interface DiagnosticsResult {
+  data class Diagnostic(val message: String, val ranges: List<TextRange>)
+
+  val diagnostics: Map<String, List<Diagnostic>>
+}
 
 class FirAnalysisResult(
   val result: FirResult,
   val reporter: BaseDiagnosticsCollector,
-) : AnalysisResult {
+) : DiagnosticsResult {
   private val sourceLocationMap by lazy { reporter.withCompilerMessageSourceLocation() }
 
   override val diagnostics by lazy {
     reporter.diagnostics.groupBy(
       keySelector = { diagnostics -> diagnostics.factoryName },
       valueTransform = { diagnostic ->
-        AnalysisResult.Diagnostic(
+        DiagnosticsResult.Diagnostic(
           message = run {
             val severity = AnalyzerWithCompilerReport.convertSeverity(diagnostic.severity)
             val renderer = RootDiagnosticRendererFactory(diagnostic)
@@ -75,34 +129,11 @@ class FirAnalysisResult(
   }
 }
 
-class K2CompilerFacade(environment: KotlinCoreEnvironment) : KotlinCompilerFacade(environment) {
+class KotlinK2Compiler private constructor(private val environment: KotlinCoreEnvironment) {
   private val project: Project get() = environment.project
   private val configuration: CompilerConfiguration get() = environment.configuration
 
-  private fun createSourceSession(
-    moduleData: FirModuleData,
-    projectSessionProvider: FirProjectSessionProvider,
-    projectEnvironment: AbstractProjectEnvironment,
-  ): FirSession =
-    FirJvmSessionFactory.createModuleBasedSession(
-      moduleData = moduleData,
-      sessionProvider = projectSessionProvider,
-      javaSourcesScope = PsiBasedProjectFileSearchScope(TopDownAnalyzerFacadeForJVM.AllJavaSourcesInProjectScope(project)),
-      projectEnvironment = projectEnvironment,
-      createIncrementalCompilationSymbolProviders = { null },
-      extensionRegistrars = FirExtensionRegistrar.getInstances(project),
-      languageVersionSettings = configuration.languageVersionSettings,
-      jvmTarget = configuration.get(JVMConfigurationKeys.JVM_TARGET) ?: error("JVM_TARGET is not specified in compiler configuration"),
-      lookupTracker = configuration.get(CommonConfigurationKeys.LOOKUP_TRACKER),
-      enumWhenTracker = configuration.get(CommonConfigurationKeys.ENUM_WHEN_TRACKER),
-      importTracker = configuration.get(CommonConfigurationKeys.IMPORT_TRACKER),
-      predefinedJavaComponents = null,
-      needRegisterJavaElementFinder = true,
-      registerExtraComponents = {},
-      init = {},
-    )
-
-  override fun analyze(file: SourceFile): FirAnalysisResult {
+  fun analyze(file: SourceFile): FirAnalysisResult {
     val moduleName = configuration.get(CommonConfigurationKeys.MODULE_NAME, "main")
 
     val projectSessionProvider = FirProjectSessionProvider()
@@ -150,14 +181,8 @@ class K2CompilerFacade(environment: KotlinCoreEnvironment) : KotlinCompilerFacad
     return FirAnalysisResult(FirResult(listOf(analysis)), reporter)
   }
 
-  private fun frontend(file: SourceFile): Fir2IrActualizedResult {
+  fun compile(file: SourceFile): Fir2IrActualizedResult {
     val analysis = analyze(file)
-
-    FirDiagnosticsCompilerResultsReporter.throwFirstErrorAsException(
-      diagnosticsCollector = analysis.reporter,
-      messageRenderer = MessageRenderer.PLAIN_FULL_PATHS,
-    )
-
     val fir2IrResult = analysis.result.convertToIrAndActualizeForJvm(
       fir2IrExtensions = JvmFir2IrExtensions(configuration = configuration, irDeserializer = JvmIrDeserializerImpl()),
       configuration = configuration,
@@ -168,5 +193,70 @@ class K2CompilerFacade(environment: KotlinCoreEnvironment) : KotlinCompilerFacad
     return fir2IrResult
   }
 
-  override fun compile(file: SourceFile): IrModuleFragment = frontend(file).irModuleFragment
+  private fun createSourceSession(
+    moduleData: FirModuleData,
+    projectSessionProvider: FirProjectSessionProvider,
+    projectEnvironment: AbstractProjectEnvironment,
+  ): FirSession =
+    FirJvmSessionFactory.createModuleBasedSession(
+      moduleData = moduleData,
+      sessionProvider = projectSessionProvider,
+      javaSourcesScope = PsiBasedProjectFileSearchScope(TopDownAnalyzerFacadeForJVM.AllJavaSourcesInProjectScope(project)),
+      projectEnvironment = projectEnvironment,
+      createIncrementalCompilationSymbolProviders = { null },
+      extensionRegistrars = FirExtensionRegistrar.getInstances(project),
+      languageVersionSettings = configuration.languageVersionSettings,
+      jvmTarget = checkNotNull(configuration.get(JVMConfigurationKeys.JVM_TARGET)) { "JVM_TARGET is not specified in compiler configuration" },
+      lookupTracker = configuration.get(CommonConfigurationKeys.LOOKUP_TRACKER),
+      enumWhenTracker = configuration.get(CommonConfigurationKeys.ENUM_WHEN_TRACKER),
+      importTracker = configuration.get(CommonConfigurationKeys.IMPORT_TRACKER),
+      predefinedJavaComponents = null,
+      needRegisterJavaElementFinder = true,
+      registerExtraComponents = {},
+      init = {},
+    )
+
+  companion object {
+    private const val TEST_MODULE_NAME = "test-module"
+
+    fun create(
+      disposable: Disposable,
+      updateConfiguration: CompilerConfiguration.() -> Unit,
+      registerExtensions: Project.(CompilerConfiguration) -> Unit,
+    ): KotlinK2Compiler {
+      val configuration = CompilerConfiguration().apply {
+        put(CommonConfigurationKeys.MODULE_NAME, TEST_MODULE_NAME)
+        put(JVMConfigurationKeys.IR, true)
+        put(CommonConfigurationKeys.USE_FIR, true)
+        put(CommonConfigurationKeys.VERIFY_IR, IrVerificationMode.ERROR)
+        put(CommonConfigurationKeys.ENABLE_IR_VISIBILITY_CHECKS, true)
+        put(JVMConfigurationKeys.JVM_TARGET, JvmTarget.JVM_17)
+        messageCollector = TestMessageCollector
+        updateConfiguration()
+      }
+
+      val environment = KotlinCoreEnvironment.createForTests(
+        parentDisposable = disposable,
+        initialConfiguration = configuration,
+        extensionConfigs = EnvironmentConfigFiles.JVM_CONFIG_FILES,
+      )
+      environment.project.registerExtensions(configuration)
+
+      return KotlinK2Compiler(environment)
+    }
+  }
+}
+
+private object TestMessageCollector : MessageCollector {
+  override fun clear() {}
+
+  override fun report(
+    severity: CompilerMessageSeverity,
+    message: String,
+    location: CompilerMessageSourceLocation?,
+  ) {
+    if (severity.isError) error(message)
+  }
+
+  override fun hasErrors(): Boolean = false
 }
